refactor(robot): extract shared review logic into lib/review.py
Deduplicate the review prompt, JSON parsing, result validation, body truncation, and template/type-key resolution that were copied between call_llm.py and tests/test_csv.py. Both now import from lib/review.py. CLI interface of call_llm.py is unchanged.
